Refactor new APIs from Server into  cluster server
