+++ src/platform/webos/PlatformManagerImpl.h
@@ -56,9 +56,6 @@ public:
 
     System::Clock::Timestamp GetStartTime() { return mStartTime; }
 
-    void HandleGeneralFault(uint32_t EventId);
-    void HandleSoftwareFault(uint32_t EventId);
-
 private:
     // ===== Methods that implement the PlatformManager abstract interface.
